lesson_10_db/requirements.txt
python-dotenv==1.0.1
loguru==0.7.2
SQLAlchemy==2.0.28
aiosqlite==0.20.0
greenlet==3.0.3
alembic==1.13.1